requirements.txt
pyTelegramBotAPI
requests~=2.32.3
Flask~=2.2.5
apache-airflow
nltk~=3.8.1
langchain~=0.2.2
openai~=1.31.2